cleanup: use standard url.JoinPath for API URL construction
