+++ pkg/provider/provider_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/stretchr/testify/assert"
 )
 
 // providerFactories are used to instantiate a provider during acceptance testing.
@@ -37,5 +38,64 @@ func testAccPreCheck(t *testing.T) {
 func testEnvIsSet(k string, t *testing.T) {
 	if v := os.Getenv(k); v == "" {
 		t.Fatalf("%[1]s must be set for acceptance tests", k)
+	}
+}
+
+func TestProviderConfigure_PathBasedURL(t *testing.T) {
+	testCases := []struct {
+		name        string
+		inputURL    string
+		expectedURL string
+	}{
+		{
+			name:        "Root path with trailing slash",
+			inputURL:    "https://api.example.com/",
+			expectedURL: "https://api.example.com/api/v3",
+		},
+		{
+			name:        "Root path without trailing slash",
+			inputURL:    "https://api.example.com",
+			expectedURL: "https://api.example.com/api/v3",
+		},
+		{
+			name:        "Single segment path with trailing slash",
+			inputURL:    "https://api.example.com/sso/",
+			expectedURL: "https://api.example.com/sso/api/v3",
+		},
+		{
+			name:        "Single segment path without trailing slash",
+			inputURL:    "https://api.example.com/sso",
+			expectedURL: "https://api.example.com/sso/api/v3",
+		},
+		{
+			name:        "Multi-segment path with trailing slash",
+			inputURL:    "https://api.example.com/auth/sso/",
+			expectedURL: "https://api.example.com/auth/sso/api/v3",
+		},
+		{
+			name:        "Multi-segment path without trailing slash",
+			inputURL:    "https://api.example.com/auth/sso",
+			expectedURL: "https://api.example.com/auth/sso/api/v3",
+		},
+		{
+			name:        "HTTP scheme with path",
+			inputURL:    "http://localhost:9000/sso/",
+			expectedURL: "http://localhost:9000/sso/api/v3",
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			p := Provider("testing", true)
+
+			_ac, diag := p.ConfigureContextFunc(t.Context(), schema.TestResourceDataRaw(t, p.Schema, map[string]any{
+				"url":      tc.inputURL,
+				"token":    "",
+				"insecure": false,
+			}))
+			assert.Nil(t, diag)
+			ac := _ac.(*APIClient)
+			assert.Equal(t, tc.expectedURL, ac.client.GetConfig().Servers[0].URL, "Server URL should be constructed correctly")
+		})
 	}
 }
